refactor: remove redundant 'getTransactionsForWallet' method
