+++ src/app/wallets/get-transactions-for-wallet.ts
@@ -20,14 +20,31 @@ export const getTransactionsForWalletId = async ({
   const wallets = WalletsRepository()
   const wallet = await wallets.findById(walletId)
   if (wallet instanceof RepositoryError) return PartialResult.err(wallet)
-  return getTransactionsForWallet(wallet)
+  return getTransactionsForWallets([wallet])
 }
 
-export const getTransactionsForWallet = async (
-  wallet: Wallet,
+export const getTransactionsForWalletIds = async (
+  walletIds: WalletId[],
 ): Promise<PartialResult<WalletTransaction[]>> => {
+  const walletsRepo = WalletsRepository()
+
+  const wallets: Wallet[] = []
+  for (const walletId of walletIds) {
+    const wallet = await walletsRepo.findById(walletId)
+    if (wallet instanceof RepositoryError) return PartialResult.err(wallet)
+
+    wallets.push(wallet)
+  }
+  return getTransactionsForWallets(wallets)
+}
+
+export const getTransactionsForWallets = async (
+  wallets: Wallet[],
+): Promise<PartialResult<WalletTransaction[]>> => {
+  const walletIds = wallets.map((wallet) => wallet.id)
+
   const ledger = LedgerService()
-  const ledgerTransactions = await ledger.getTransactionsByWalletId(wallet.id)
+  const ledgerTransactions = await ledger.getTransactionsByWalletIds(walletIds)
   if (ledgerTransactions instanceof LedgerError)
     return PartialResult.err(ledgerTransactions)
 
@@ -47,13 +64,22 @@ export const getTransactionsForWallet = async (
     return PartialResult.partial(confirmedHistory.transactions, onChainTxs)
   }
 
-  const addresses = wallet.onChainAddresses()
+  const addresses: OnChainAddress[] = []
+  const addressesByWalletId: { walletId: OnChainAddress[] } = {} as {
+    walletId: OnChainAddress[]
+  }
+  for (const wallet of wallets) {
+    const walletAddresses = wallet.onChainAddresses()
+    addressesByWalletId[wallet.id] = walletAddresses
+    addresses.push(...walletAddresses)
+  }
+
   const filter = TxFilter({
     confirmationsLessThan: ONCHAIN_MIN_CONFIRMATIONS,
     addresses,
   })
 
-  const pendingTxs = wrapToRunInSpan({
+  const pendingIncoming = wrapToRunInSpan({
     namespace: `domain.bitcoin`,
     fn: () => filter.apply(onChainTxs),
   })()
@@ -65,8 +91,8 @@ export const getTransactionsForWallet = async (
 
   return PartialResult.ok(
     confirmedHistory.addPendingIncoming({
-      pendingIncoming: pendingTxs,
-      addressesByWalletId: { [wallet.id]: addresses },
+      pendingIncoming,
+      addressesByWalletId,
       displayCurrencyPerSat: price,
     }).transactions,
   )
